==> 1 - SSB ban - NHANES.Rmd <==
---
title: "SSB ban - NHANES"
author: "Sanjay Basu"
date: "5/28/2019"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
install.packages('dplyr', repos='http://cran.us.r-project.org')
install.packages('tidyr', repos='http://cran.us.r-project.org')
install.packages('purrr', repos='http://cran.us.r-project.org')
install.packages('nhanesA', repos='http://cran.us.r-project.org')
install.packages('furrr', repos='http://cran.us.r-project.org')
install.packages('mice', repos='http://cran.us.r-project.org')
install.packages('survey', repos='http://cran.us.r-project.org')

library(dplyr)
library(tidyr)
library(purrr)
library(nhanesA)
library(furrr)
library(mice)
library(survey)
library(readr)
```

## NHANES import

extract and define needed variables from NHANES 2011-2016
note the CKD-EPI equation is used to define CKD, http://nephron.com/epi_equation ; our justification for this decision was the results from: https://jamanetwork.com/journals/jama/fullarticle/1151529

```{r NHANES 2011-2012}
demo_g = nhanes('DEMO_G') %>%
  mutate(age = RIDAGEYR,
         female = (RIAGENDR==2),
         white = (RIDRETH3==3),
         black = (RIDRETH3==4),
         asian = (RIDRETH3==6),
         other = ((RIDRETH3!=3)&(RIDRETH3!=4)&(RIDRETH3!=6)),
         hisp = (RIDRETH3<=2),
         intweight = WTINT2YR/3,
         mecweight = WTMEC2YR/3,
         psu = SDMVPSU,
         strata = SDMVSTRA) %>%
  select(SEQN, age, female, white, black, asian, other, hisp, intweight, mecweight, psu, strata)
hiq_g = nhanes('HIQ_G') %>%
  mutate(privins = (HIQ031A==14),
         privins = replace_na(privins, 0)) %>%
  select(SEQN, privins)
ocq_g = nhanes('OCQ_G') %>%
  filter((OCD231 ==16)&(OCD150 ==1)) %>%
  mutate(mgmt = (OCD241<=7),
         blcol = (((OCD241>=12)&(OCD241<=15))|((OCD241>=18)&(OCD241<=22))),
         whcol = (((OCD241>=8)&(OCD241<=11))|((OCD241>=16)&(OCD241<=17)))) %>%
  select(SEQN, mgmt, blcol, whcol)
bmx_g = nhanes('BMX_G') %>%
  mutate(bmi = BMXBMI,
         wtkg = BMXWT,
         htm = BMXHT/100) %>%
  select(SEQN, bmi, wtkg, htm)
mcq_g = nhanes('MCQ_G') %>%
  mutate(chd = (MCQ160B==1)|(MCQ160C==1)|(MCQ160D==1)|(MCQ160E==1),
         str = (MCQ160F==1)) %>%
  select(SEQN, chd, str)
rxq_drug = nhanes('RXQ_DRUG') 
rxq_rx_g = nhanes('RXQ_RX_G') %>%
  left_join(rxq_drug, by = "RXDDRGID") %>%
  filter(RXDDCN1B=="ANTIDIABETIC AGENTS") %>%
  group_by(SEQN) %>%
  mutate(dmmed = 1) %>%
  select(SEQN,dmmed)
ghb_g = nhanes('GHB_G') %>%
  select(SEQN, LBXGH)
glu_g = nhanes('GLU_G') %>%
  mutate(fastweight = WTSAF2YR/3) %>%
  select(SEQN, LBXGLU,fastweight)
ogtt_g = nhanes('OGTT_G') %>%
  select(SEQN, LBXGLT)
diq_g = nhanes('DIQ_G') %>%
  full_join(ghb_g, by="SEQN") %>%
  full_join(glu_g, by="SEQN") %>%
  full_join(ogtt_g, by="SEQN") %>%
  full_join(rxq_rx_g, by="SEQN") %>%
  mutate(dm = ((LBXGH>=6.5)|(LBXGLU>=126)|(LBXGLT>=200)|(DIQ010==1)|(dmmed==1)),
         dm = replace_na(dm, 0)) %>%
  select(SEQN,dm,fastweight)
biopro_g = nhanes('BIOPRO_G') %>%
  select(SEQN, LBXSCR)
kiq_u_g = nhanes('KIQ_U_G') %>%
  full_join(biopro_g, by="SEQN") %>%
  full_join(demo_g, by="SEQN") %>%
  mutate(scr = LBXSCR,
         kappa = 0.7*female + 0.9 *(1-female),
         alpha = -0.329*female -0.411*(1-female),
         sk = scr/kappa,
         minsk = sk*(sk<1)+1*(sk>=1),
         maxsk = sk*(sk>=1)+1*(sk<1),
         egfr = 141 *(minsk^alpha) *  
           (maxsk^(-1.209))*
           (0.993^age)*
           (1.018*female+1*(1-female))*
           (1.159*black+1*(1-black)),
         egfr_mdrd = 186 * (scr^(-1.154))*
           (age^(-0.203))*
          (1.212*black+1*(1-black))*
          (0.742*female+1*(1-female)),
         ckd = (egfr<90),
         ckd_mdrd = (egfr_mdrd<90)) %>%
  select(SEQN, ckd, ckd_mdrd)
ohxref_g = nhanes('OHXREF_G') %>%
  mutate(dental = (OHAROCDT==1)&(OHAROCGP==1)) %>%
  select(SEQN, dental)
```

```{r NHANES 2013-2014}
demo_h = nhanes('DEMO_H') %>%
  mutate(age = RIDAGEYR,
         female = (RIAGENDR==2),
         white = (RIDRETH3==3),
         black = (RIDRETH3==4),
         asian = (RIDRETH3==6),
         other = ((RIDRETH3!=3)&(RIDRETH3!=4)&(RIDRETH3!=6)),
         hisp = (RIDRETH3<=2),
         intweight = WTINT2YR/3,
         mecweight = WTMEC2YR/3,
         psu = SDMVPSU,
         strata = SDMVSTRA) %>%
  select(SEQN, age, female, white, black, asian, other, hisp, intweight, mecweight, psu, strata)
hiq_h = nhanes('HIQ_H') %>%
  mutate(privins = (HIQ031A==14),
         privins = replace_na(privins, 0)) %>%
  select(SEQN, privins)
ocq_h = nhanes('OCQ_H') %>%
  filter((OCQ260 ==1)&(OCD150 ==1)) %>%
  mutate(mgmt = NA,
         blcol = NA,
         whcol = NA) %>%
  select(SEQN, mgmt, blcol, whcol)
bmx_h = nhanes('BMX_H') %>%
  mutate(bmi = BMXBMI,
         wtkg = BMXWT,
         htm = BMXHT/100) %>%
  select(SEQN, bmi, wtkg, htm)
mcq_h = nhanes('MCQ_H') %>%
  mutate(chd = (MCQ160B==1)|(MCQ160C==1)|(MCQ160D==1)|(MCQ160E==1),
         str = (MCQ160F==1)) %>%
  select(SEQN, chd, str)
rxq_drug = nhanes('RXQ_DRUG') 
rxq_rx_h = nhanes('RXQ_RX_H') %>%
  left_join(rxq_drug, by = "RXDDRGID") %>%
  filter(RXDDCN1B=="ANTIDIABETIC AGENTS") %>%
  group_by(SEQN) %>%
  mutate(dmmed = 1) %>%
  select(SEQN,dmmed)
ghb_h = nhanes('GHB_H') %>%
  select(SEQN, LBXGH)
glu_h = nhanes('GLU_H') %>%
  mutate(fastweight = WTSAF2YR/3) %>%
  select(SEQN, LBXGLU,fastweight)
ogtt_h = nhanes('OGTT_H') %>%
  select(SEQN, LBXGLT)
diq_h = nhanes('DIQ_H') %>%
  full_join(ghb_h, by="SEQN") %>%
  full_join(glu_h, by="SEQN") %>%
  full_join(ogtt_h, by="SEQN") %>%
  full_join(rxq_rx_h, by="SEQN") %>%
  mutate(dm = ((LBXGH>=6.5)|(LBXGLU>=126)|(LBXGLT>=200)|(DIQ010==1)|(dmmed==1)),
         dm = replace_na(dm, 0)) %>%
  select(SEQN,dm,fastweight)
biopro_h = nhanes('BIOPRO_H') %>%
  select(SEQN, LBXSCR)
kiq_u_h = nhanes('KIQ_U_H') %>%
  full_join(biopro_h, by="SEQN") %>%
  full_join(demo_h, by="SEQN") %>%
  mutate(scr = LBXSCR,
         kappa = 0.7*female + 0.9 *(1-female),
         alpha = -0.329*female -0.411*(1-female),
         sk = scr/kappa,
         minsk= sk*(sk<1)+1*(sk>=1),
         maxsk = sk*(sk>=1)+1*(sk<1),
         egfr = 141 *(minsk^alpha) *
           (maxsk^(-1.209))*
           (0.993^age)*
           (1.018*female+1*(1-female))*
           (1.159*black+1*(1-black)),
         egfr_mdrd = 186 * (scr^(-1.154))*
           (age^(-0.203))*
          (1.212*black+1*(1-black))*
          (0.742*female+1*(1-female)),
         ckd = (egfr<90),
         ckd_mdrd = (egfr_mdrd<90)) %>%
  select(SEQN, ckd, ckd_mdrd)
ohxref_h = nhanes('OHXREF_H') %>%
  mutate(dental = (OHAROCDT==1)&(OHAROCGP==1)) %>%
  select(SEQN, dental)
```

```{r NHANES 2015-2016}
demo_i = nhanes('DEMO_I') %>%
  mutate(age = RIDAGEYR,
         female = (RIAGENDR==2),
         white = (RIDRETH3==3),
         black = (RIDRETH3==4),
         asian = (RIDRETH3==6),
         other = ((RIDRETH3!=3)&(RIDRETH3!=4)&(RIDRETH3!=6)),
         hisp = (RIDRETH3<=2),
         intweight = WTINT2YR/3,
         mecweight = WTMEC2YR/3,
         psu = SDMVPSU,
         strata = SDMVSTRA) %>%
  select(SEQN, age, female, white, black, asian, other, hisp, intweight, mecweight, psu, strata)
hiq_i = nhanes('HIQ_I') %>%
  mutate(privins = (HIQ031A==14),
         privins = replace_na(privins, 0)) %>%
  select(SEQN, privins)
ocq_i = nhanes('OCQ_I') %>%
  filter((OCQ260 ==1)&(OCD150 ==1)) %>%
  mutate(mgmt = NA,
         blcol = NA,
         whcol = NA) %>%
  select(SEQN, mgmt, blcol, whcol)
bmx_i = nhanes('BMX_I') %>%
  mutate(bmi = BMXBMI,
         wtkg = BMXWT,
         htm = BMXHT/100) %>%
  select(SEQN, bmi, wtkg, htm)
mcq_i = nhanes('MCQ_I') %>%
  mutate(chd = (MCQ160B==1)|(MCQ160C==1)|(MCQ160D==1)|(MCQ160E==1),
         str = (MCQ160F==1)) %>%
  select(SEQN, chd, str)
rxq_drug = nhanes('RXQ_DRUG') 
rxq_rx_i = nhanes('RXQ_RX_I') %>%
  left_join(rxq_drug, by = "RXDDRGID") %>%
  filter(RXDDCN1B=="ANTIDIABETIC AGENTS") %>%
  group_by(SEQN) %>%
  mutate(dmmed = 1) %>%
  select(SEQN,dmmed)
ghb_i = nhanes('GHB_I') %>%
  select(SEQN, LBXGH)
glu_i = nhanes('GLU_I') %>%
  mutate(fastweight = WTSAF2YR/3) %>%
  select(SEQN, LBXGLU,fastweight)
ogtt_i = nhanes('OGTT_I') %>%
  select(SEQN, LBXGLT)
diq_i = nhanes('DIQ_I') %>%
  full_join(ghb_i, by="SEQN") %>%
  full_join(glu_i, by="SEQN") %>%
  full_join(ogtt_i, by="SEQN") %>%
  full_join(rxq_rx_i, by="SEQN") %>%
  mutate(dm = ((LBXGH>=6.5)|(LBXGLU>=126)|(LBXGLT>=200)|(DIQ010==1)|(dmmed==1)),
         dm = replace_na(dm, 0)) %>%
  select(SEQN,dm,fastweight)
biopro_i = nhanes('BIOPRO_I') %>%
  select(SEQN, LBXSCR)
kiq_u_i = nhanes('KIQ_U_I') %>%
  full_join(biopro_i, by="SEQN") %>%
  full_join(demo_i, by="SEQN") %>%
  mutate(scr = LBXSCR,
         kappa = 0.7*female + 0.9 *(1-female),
         alpha = -0.329*female -0.411*(1-female),
         sk = scr/kappa,
         minsk = sk*(sk<1)+1*(sk>=1),
         maxsk = sk*(sk>=1)+1*(sk<1),
         egfr = 141 *(minsk^alpha) *
           (maxsk^(-1.209))*
           (0.993^age)*
           (1.018*female+1*(1-female))*
           (1.159*black+1*(1-black)),
         egfr_mdrd = 186 * (scr^(-1.154))*
           (age^(-0.203))*
          (1.212*black+1*(1-black))*
          (0.742*female+1*(1-female)),
         ckd = (egfr<90),
         ckd_mdrd = (egfr_mdrd<90)) %>%
  select(SEQN, ckd, ckd_mdrd)
ohxref_i = nhanes('OHXREF_I') %>%
  mutate(dental = (OHAROCDT==1)&(OHAROCGP==1)) %>%
  select(SEQN, dental)
```

## NHANES join

join and limit to employed pop by occ class, starting with ocq and left joins to limit to those employed in priv sec

```{r NHANES data joins}

nhanes11 = ocq_g %>%
  left_join(demo_g, by="SEQN") %>%
  left_join(hiq_g, by="SEQN") %>%
  left_join(bmx_g, by="SEQN") %>%
  left_join(mcq_g, by="SEQN") %>%
  left_join(diq_g, by="SEQN") %>%
  left_join(kiq_u_g, by="SEQN") %>%
  left_join(ohxref_g, by="SEQN")

nhanes13 = ocq_h %>%
  left_join(demo_h, by="SEQN") %>%
  left_join(hiq_h, by="SEQN") %>%
  left_join(bmx_h, by="SEQN") %>%
  left_join(mcq_h, by="SEQN") %>%
  left_join(diq_h, by="SEQN") %>%
  left_join(kiq_u_h, by="SEQN") %>%
  left_join(ohxref_h, by="SEQN")  

nhanes15 = ocq_i %>%
  left_join(demo_i, by="SEQN") %>%
  left_join(hiq_i, by="SEQN") %>%
  left_join(bmx_i, by="SEQN") %>%
  left_join(mcq_i, by="SEQN") %>%
  left_join(diq_i, by="SEQN") %>%
  left_join(kiq_u_i, by="SEQN") %>%
  left_join(ohxref_i, by="SEQN") %>%
  mutate(obese = (bmi>=30))

nhanes = bind_rows(list(nhanes11, nhanes13, nhanes15)) 
nhanes$mgmt <- as.factor(nhanes$mgmt)
nhanes$blcol <- as.factor(nhanes$blcol)
nhanes$whcol <- as.factor(nhanes$whcol)
nhanes$female <- as.factor(nhanes$female)
nhanes$white <- as.factor(nhanes$white)
nhanes$black <- as.factor(nhanes$black)
nhanes$asian <- as.factor(nhanes$asian)
nhanes$other <- as.factor(nhanes$other)
nhanes$hisp <- as.factor(nhanes$hisp)
nhanes$privins <- as.factor(nhanes$privins)
nhanes$chd <- as.factor(nhanes$chd)
nhanes$str <- as.factor(nhanes$str)
nhanes$dm <- as.factor(nhanes$dm)
nhanes$ckd <- as.factor(nhanes$ckd)
nhanes$ckd_mdrd <- as.factor(nhanes$ckd_mdrd)
nhanes$dental <- as.factor(nhanes$dental)
nhanes$obese <- as.factor(nhanes$obese)

nhanesDesign <- svydesign(id      = ~psu,
                          strata  = ~strata,
                          weights = ~intweight,
                          nest    = TRUE,
                          data    = nhanes)
nhanes_mec = nhanes %>%
  filter(mecweight!='NA')
nhanesDesign_mec <- svydesign(id      = ~psu,
                          strata  = ~strata,
                          weights = ~mecweight,
                          nest    = TRUE,
                          data    = nhanes_mec)
nhanes_lab = nhanes %>%
  filter(fastweight!='NA')
nhanesDesign_lab <- svydesign(id      = ~psu,
                          strata  = ~strata,
                          weights = ~fastweight,
                          nest    = TRUE,
                          data    = nhanes_lab)

summary(nhanesDesign)
summary(nhanesDesign_mec)
summary(nhanesDesign_lab)

```
## weighted stats
```{r weighted stats}
svymean(~age, nhanesDesign, na.rm = TRUE)
svyquantile(~age, nhanesDesign, quantiles=c(.25,.75),na.rm = TRUE)
svymean(~female, nhanesDesign, na.rm = TRUE)
svymean(~white, nhanesDesign, na.rm = TRUE)
svymean(~black, nhanesDesign, na.rm = TRUE)
svymean(~hisp, nhanesDesign, na.rm = TRUE)
svymean(~mgmt, nhanesDesign, na.rm = TRUE)
svymean(~whcol, nhanesDesign, na.rm = TRUE)
svymean(~blcol, nhanesDesign, na.rm = TRUE)
svymean(~hisp, nhanesDesign, na.rm = TRUE)
svymean(~obese, nhanesDesign_mec, na.rm = TRUE)
svymean(~chd, nhanesDesign_mec, na.rm = TRUE)
svymean(~str, nhanesDesign_mec, na.rm = TRUE)
svymean(~dm, nhanesDesign_mec, na.rm = TRUE)
svymean(~ckd, nhanesDesign_lab, na.rm = TRUE)
svymean(~dental, nhanesDesign_mec, na.rm = TRUE)

```
## Imputation

10-fold multiple imputation w/ chained equations

```{r imputation}
set.seed(123)
plan(multiprocess)
start = Sys.time()
imp_nhanes = future_map(rep(1, 10), ~mice(data = nhanes, m = ., maxit = 100, imputationWeights = nhanes$mecweight, printFlag = FALSE))
end = Sys.time() - start
print(end)
imp_nhanes_complete = map(imp_nhanes, mice::complete)
imp_nhanes = map2(.x = seq(1,10), .y = imp_nhanes_complete, ~mutate(.y, imp_id = as.character(.x)))
imp_nhanes = bind_rows(imp_nhanes)
summary(imp_nhanes)
write_csv(imp_nhanes, path = "~/imp_nhanes.csv")

```
