==> 3 - SSB ban - MEPS.Rmd <==
---
title: "3 - SSB ban - MEPS"
author: "Sanjay Basu"
date: "5/30/2019"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
  package_names <- c("survey","dplyr","foreign","devtools")
  lapply(package_names, function(x) if(!x %in% installed.packages()) install.packages(x, repos='http://cran.us.r-project.org'))
  lapply(package_names, require, character.only=T)

  install_github("e-mitchell/meps_r_pkg/MEPS")
  library(MEPS)

  options(survey.lonely.psu="adjust")
```

## MEPS 2015
Based on: https://meps.ahrq.gov/mepstrends/hc_cond/

```{r import MEPS}
  temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h181ssp.zip",temp)
  FYC <- read.xport(unzip(temp))
  unlink(temp)
  year <- 2015

  if(year <= 2001) FYC <- FYC %>% mutate(VARPSU = VARPSU15, VARSTR=VARSTR15)
  if(year <= 1998) FYC <- FYC %>% rename(PERWT15F = WTDPER15)
  if(year == 1996) FYC <- FYC %>% mutate(AGE42X = AGE2X, AGE31X = AGE1X)

  FYC <- FYC %>%
    mutate_at(vars(starts_with("AGE")),funs(replace(., .< 0, NA))) %>%
    mutate(AGELAST = coalesce(AGE15X, AGE42X, AGE31X))

  FYC$ind = 1  

# Employment Status
  if(year == 1996)
    FYC <- FYC %>% mutate(EMPST53 = EMPST96, EMPST42 = EMPST2, EMPST31 = EMPST1)

  FYC <- FYC %>%
    mutate_at(vars(EMPST53, EMPST42, EMPST31), funs(replace(., .< 0, NA))) %>%
    mutate(employ_last = coalesce(EMPST53, EMPST42, EMPST31))

  FYC <- FYC %>% mutate(
    employed = 1*(employ_last==1) + 2*(employ_last > 1),
    employed = replace(employed, is.na(employed) & AGELAST < 16, 9),
    employed = recode_factor(employed, .default = "Missing", .missing = "Missing", 
      "1" = "Employed",
      "2" = "Not employed",
      "9" = "Inapplicable (age < 16)"))

# Keep only needed variables from FYC
  FYCsub <- FYC %>% select(employed,ind, DUPERSID, PERWT15F, VARSTR, VARPSU)
```

# Load event files

```{r Load event files}
temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178assp.zip",temp)
  RX <- read.xport(unzip(temp))
  unlink(temp)
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178bssp.zip",temp)
  DVT <- read.xport(unzip(temp))
  unlink(temp)
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178dssp.zip",temp)
  IPT <- read.xport(unzip(temp))
  unlink(temp)
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178dssp.zip",temp)
  IPT <- read.xport(unzip(temp))
  unlink(temp)  
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178essp.zip",temp)
  ERT <- read.xport(unzip(temp))
  unlink(temp)  
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178fssp.zip",temp)
  OPT <- read.xport(unzip(temp))
  unlink(temp)    
  unlink(temp)  
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178gssp.zip",temp)
  OBV <- read.xport(unzip(temp))
  unlink(temp)    
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178hssp.zip",temp)
  HHT <- read.xport(unzip(temp))
  unlink(temp)    
  

```

# Define sub-levels for office-based and outpatient
To compute estimates for these sub-events, replace 'event' with 'event_v2X'
in the 'svyby' statement below, when applicable

```{r Define sub-levels}
OBV <- OBV %>%
    mutate(event_v2X = recode_factor(
      SEEDOC, .default = 'Missing', .missing = "Missing", '1' = 'OBD', '2' = 'OBO'))

  OPT <- OPT %>%
    mutate(event_v2X = recode_factor(
      SEEDOC, .default = 'Missing', .missing = "Missing", '1' = 'OPY', '2' = 'OPZ'))
```

# Sum RX purchases for each event

```{r Sum RX}
RX <- RX %>%
    rename(EVNTIDX = LINKIDX) %>%
    group_by(DUPERSID,EVNTIDX) %>%
    summarise_at(vars(RXSF15X:RXXP15X),sum) %>%
    ungroup
```

# Stack events (dental visits and other medical not collected for events)

```{r Stack events}
stacked_events <- stack_events(RX, IPT, ERT, OPT, OBV, HHT, keep.vars = c('SEEDOC','event_v2X'))

  stacked_events <- stacked_events %>%
    mutate(event = data,
           PR15X = PV15X + TR15X,
           OZ15X = OF15X + SL15X + OT15X + OR15X + OU15X + WC15X + VA15X)
```

# Read in event-condition linking file

```{r Read in event-condition}
     temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h178if1ssp.zip",temp)
  clink1 <- read.xport(unzip(temp))
  unlink(temp)  
  clink1 = clink1 %>%
    select(DUPERSID,CONDIDX,EVNTIDX)
```

# Read in conditions file and merge with condition_codes, link file

```{r Read in conditions file}
temp <- tempfile()
  download.file("https://meps.ahrq.gov/mepsweb/data_files/pufs/h180ssp.zip",temp)
  cond <- read.xport(unzip(temp))
  unlink(temp)      
  
  cond <- cond %>%
    select(DUPERSID, CONDIDX, CCCODEX) %>%
    mutate(CCS_Codes = as.numeric(as.character(CCCODEX))) %>%
    left_join(condition_codes, by = "CCS_Codes") %>%
    full_join(clink1, by = c("DUPERSID", "CONDIDX")) %>%
    distinct(DUPERSID, EVNTIDX, Condition, .keep_all=T)
```

# Merge events with conditions-link file and FYCsub

```{r Merge events with conditions-link file and FYCsub}
all_events <- full_join(stacked_events, cond, by=c("DUPERSID","EVNTIDX")) %>%
    filter(!is.na(Condition),XP15X >= 0) %>%
    mutate(count = 1) %>%
    full_join(FYCsub, by = "DUPERSID")
```

# Sum by person, condition, across event

```{r Sum by person, condition, across event}
all_pers <- all_events %>%
  group_by(employed,ind, DUPERSID, VARSTR, VARPSU, PERWT15F, Condition, count) %>%
  summarize_at(vars(SF15X, PR15X, MR15X, MD15X, OZ15X, XP15X),sum) %>% ungroup

PERSdsgn <- svydesign(
  id = ~VARPSU,
  strata = ~VARSTR,
  weights = ~PERWT15F,
  data = all_pers,
  nest = TRUE)

results <- svyby(~XP15X, by = ~Condition + employed, FUN = svymean, design = PERSdsgn)
print(results)
```
