==> 5 - SSB ban - postprocessing.Rmd <==
---
title: "5 - SSB ban - postprocessing"
author: "Sanjay Basu"
date: "6/10/2019"
output: pdf_document
---

```{r setup, include=FALSE}
rm(list=ls())
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(matrixStats)
```


## CEA of SSB ban

```{r cea}

load('SSBbansim.RData')

(mean(prev_obese[,1:10])-mean(prev_obese_post[,1:10]))*100
(mean(prev_chd[,1:10])-mean(prev_chd_post[,1:10]))*100
(mean(prev_str[,1:10])-mean(prev_str_post[,1:10]))*100
(mean(prev_dm[,1:10])-mean(prev_dm_post[,1:10]))*100
(mean(prev_ckd[,1:10])-mean(prev_ckd_post[,1:10]))*100
(mean(prev_den[,1:10])-mean(prev_den_post[,1:10]))*100
(mean(prev_obese[,1:10])-mean(prev_obese_post[,1:10]))*100/mean(prev_obese[,1:10])
(mean(prev_chd[,1:10])-mean(prev_chd_post[,1:10]))*100/mean(prev_chd[,1:10])
(mean(prev_str[,1:10])-mean(prev_str_post[,1:10]))*100/mean(prev_str[,1:10])
(mean(prev_dm[,1:10])-mean(prev_dm_post[,1:10]))*100/mean(prev_dm[,1:10])
(mean(prev_ckd[,1:10])-mean(prev_ckd_post[,1:10]))*100/mean(prev_ckd[,1:10])
(mean(prev_den[,1:10])-mean(prev_den_post[,1:10]))*100/mean(prev_den[,1:10])


summary(tot_costs_10)
quantile(tot_costs_10,c(.025,.975))

sum(((costs_obese+prod_obese)*prev_obese[,1:10]))/sum(costs[,1:10])
sum(((costs_chd+prod_chd)*prev_chd[,1:10]))/sum(costs[,1:10])
sum(((costs_str+prod_str)*prev_str[,1:10]))/sum(costs[,1:10])
sum(((costs_dm+prod_dm)*prev_dm[,1:10]))/sum(costs[,1:10])
sum(((costs_ckd+prod_ckd)*prev_ckd[,1:10]))/sum(costs[,1:10])
sum(((costs_den+prod_den)*prev_den[,1:10]))/sum(costs[,1:10])

1.391106/(1.391106+0.2195245+0.2164296+0.8643616+2.524564+1.10143)
0.2195245/(1.391106+0.2195245+0.2164296+0.8643616+2.524564+1.10143)
0.2164296/(1.391106+0.2195245+0.2164296+0.8643616+2.524564+1.10143)
0.8643616/(1.391106+0.2195245+0.2164296+0.8643616+2.524564+1.10143)
2.524564/(1.391106+0.2195245+0.2164296+0.8643616+2.524564+1.10143)
1.10143/(1.391106+0.2195245+0.2164296+0.8643616+2.524564+1.10143)


summary(tot_costs_lt)
quantile(tot_costs_lt,c(.025,.975))

sum(((costs_obese+prod_obese)*prev_obese))/sum(costs)
sum(((costs_chd+prod_chd)*prev_chd))/sum(costs)
sum(((costs_str+prod_str)*prev_str))/sum(costs)
sum(((costs_dm+prod_dm)*prev_dm))/sum(costs)
sum(((costs_ckd+prod_ckd)*prev_ckd))/sum(costs)
sum(((costs_den+prod_den)*prev_den))/sum(costs)
3.552227/(3.552227+4.163938+6.493774+3.365163+10.39669+1.178865)
4.163938/(3.552227+4.163938+6.493774+3.365163+10.39669+1.178865)
6.493774/(3.552227+4.163938+6.493774+3.365163+10.39669+1.178865)
3.365163/(3.552227+4.163938+6.493774+3.365163+10.39669+1.178865)
10.39669/(3.552227+4.163938+6.493774+3.365163+10.39669+1.178865)
1.178865/(3.552227+4.163938+6.493774+3.365163+10.39669+1.178865)

summary(tot_qalys_10)
quantile(tot_qalys_10,c(.025,.975))

summary(tot_qalys_lt)
quantile(tot_qalys_lt,c(.025,.975))



(mean((imp_nhanes$wtkg-delta_wtkg)/(imp_nhanes$htm^2)>30)-mean(imp_nhanes$bmi>30))/3/mean(imp_nhanes$bmi>30)*100
(1-delta_chd)*100
(1-delta_str)*100
(1-delta_dm)*100
(1-delta_ckd)*100
(1-delta_den)*100
(1-delta_mort)*100


deltaqual_10 = (tot_qalys_10_post-tot_qalys_10)
deltaqual_10[deltaqual_10<0]=mean(deltaqual_10)
summary(deltaqual_10)
quantile(deltaqual_10,c(.025,.975))
sum(((1*(1-prev_obese_post[,1:10])+comp_util_obese*prev_obese_post[,1:10])*alive_post[,1:10])-((1*(1-prev_obese[,1:10])+comp_util_obese*prev_obese[,1:10])*alive[,1:10]))/sum(tot_qalys_10_post-tot_qalys_10)*100 
sum(((1*(1-prev_chd_post[,1:10])+comp_util_chd*prev_chd_post[,1:10])*alive_post[,1:10])-((1*(1-prev_chd[,1:10])+comp_util_chd*prev_chd[,1:10])*alive[,1:10]))/sum(tot_qalys_10_post-tot_qalys_10)*100
sum(((1*(1-prev_str_post[,1:10])+comp_util_str*prev_str_post[,1:10])*alive_post[,1:10])-((1*(1-prev_str[,1:10])+comp_util_str*prev_str[,1:10])*alive[,1:10]))/sum(tot_qalys_10_post-tot_qalys_10)*100
sum(((1*(1-prev_dm_post[,1:10])+comp_util_dm*prev_dm_post[,1:10])*alive_post[,1:10])-((1*(1-prev_dm[,1:10])+comp_util_dm*prev_dm[,1:10])*alive[,1:10]))/sum(tot_qalys_10_post-tot_qalys_10)*100
sum(((1*(1-prev_ckd_post[,1:10])+comp_util_ckd*prev_ckd_post[,1:10])*alive_post[,1:10])-((1*(1-prev_ckd[,1:10])+comp_util_ckd*prev_ckd[,1:10])*alive[,1:10]))/sum(tot_qalys_10_post-tot_qalys_10)*100
sum(((1*(1-prev_den_post[,1:10])+comp_util_den*prev_den_post[,1:10])*alive_post[,1:10])-((1*(1-prev_den[,1:10])+comp_util_den*prev_den[,1:10])*alive[,1:10]))/sum(tot_qalys_10_post-tot_qalys_10)*100
sum(alive_post[,1:10]-alive[,1:10])/sum(tot_qalys_10_post-tot_qalys_10)*100 
454.624/(454.624+185.8462+189.4105+190.9328+191.5768+205.4616+191.7668)
185.8462/(454.624+185.8462+189.4105+190.9328+191.5768+205.4616+191.7668)
189.4105/(454.624+185.8462+189.4105+190.9328+191.5768+205.4616+191.7668)
190.9328/(454.624+185.8462+189.4105+190.9328+191.5768+205.4616+191.7668)
191.5768/(454.624+185.8462+189.4105+190.9328+191.5768+205.4616+191.7668)
205.4616/(454.624+185.8462+189.4105+190.9328+191.5768+205.4616+191.7668)
191.7668/(454.624+185.8462+189.4105+190.9328+191.5768+205.4616+191.7668)



deltaqual_lt = (tot_qalys_lt_post-tot_qalys_lt)
deltaqual_lt[deltaqual_lt<0]=mean(deltaqual_lt)
summary(deltaqual_lt)
quantile(deltaqual_lt,c(.025,.975))
sum(((1*(1-prev_obese_post)+comp_util_obese*prev_obese_post)*alive_post)-((1*(1-prev_obese)+comp_util_obese*prev_obese)*alive))/sum(tot_qalys_lt_post-tot_qalys_lt)*100
sum(((1*(1-prev_chd_post)+comp_util_chd*prev_chd_post)*alive_post)-((1*(1-prev_chd)+comp_util_chd*prev_chd)*alive))/sum(tot_qalys_lt_post-tot_qalys_lt)*100
sum(((1*(1-prev_str_post)+comp_util_str*prev_str_post)*alive_post)-((1*(1-prev_str)+comp_util_str*prev_str)*alive))/sum(tot_qalys_lt_post-tot_qalys_lt)*100
sum(((1*(1-prev_dm_post)+comp_util_dm*prev_dm_post)*alive_post)-((1*(1-prev_dm)+comp_util_dm*prev_dm)*alive))/sum(tot_qalys_lt_post-tot_qalys_lt)*100
sum(((1*(1-prev_ckd_post)+comp_util_ckd*prev_ckd_post)*alive_post)-((1*(1-prev_ckd)+comp_util_ckd*prev_ckd)*alive))/sum(tot_qalys_lt_post-tot_qalys_lt)*100
sum(((1*(1-prev_den_post)+comp_util_den*prev_den_post)*alive_post)-((1*(1-prev_den)+comp_util_den*prev_den)*alive))/sum(tot_qalys_lt_post-tot_qalys_lt)*100
sum(alive_post-alive)/sum(tot_qalys_lt_post-tot_qalys_lt)*100
352.9331/(352.9331+285.7486+273.3949+294.3408+298.0191+300.6823+298.2317)
285.7486/(352.9331+285.7486+273.3949+294.3408+298.0191+300.6823+298.2317)
273.3949/(352.9331+285.7486+273.3949+294.3408+298.0191+300.6823+298.2317)
294.3408/(352.9331+285.7486+273.3949+294.3408+298.0191+300.6823+298.2317)
298.0191/(352.9331+285.7486+273.3949+294.3408+298.0191+300.6823+298.2317)
300.6823/(352.9331+285.7486+273.3949+294.3408+298.0191+300.6823+298.2317)
298.2317/(352.9331+285.7486+273.3949+294.3408+298.0191+300.6823+298.2317)

summary(tot_costs_10_post-tot_costs_10)
quantile(tot_costs_10_post-tot_costs_10,c(.025,.975))
sum(((costs_obese+prod_obese)*prev_obese_post[,1:10])*alive_post[,1:10]-((costs_obese+prod_obese)*prev_obese[,1:10])*alive[,1:10])/sum(tot_costs_10_post-tot_costs_10)*100
sum(((costs_chd+prod_chd)*prev_chd_post[,1:10])*alive_post[,1:10]-((costs_chd+prod_chd)*prev_chd[,1:10])*alive[,1:10])/sum(tot_costs_10_post-tot_costs_10)*100
sum(((costs_str+prod_str)*prev_str_post)[,1:10]*alive_post[,1:10]-((costs_str+prod_str)*prev_str[,1:10])*alive[,1:10])/sum(tot_costs_10_post-tot_costs_10)*100
sum(((costs_dm+prod_dm)*prev_dm_post[,1:10])*alive_post[,1:10]-((costs_dm+prod_dm)*prev_dm[,1:10])*alive[,1:10])/sum(tot_costs_10_post-tot_costs_10)*100
sum(((costs_ckd+prod_ckd)*prev_ckd_post[,1:10])*alive_post[,1:10]-((costs_ckd+prod_ckd)*prev_ckd[,1:10])*alive[,1:10])/sum(tot_costs_10_post-tot_costs_10)*100
sum(((costs_den+prod_den)*prev_den_post[,1:10])*alive_post[,1:10]-((costs_den+prod_den)*prev_den[,1:10])*alive[,1:10])/sum(tot_costs_10_post-tot_costs_10)*100
8.712809/(8.712809+0.1114585+0.03197762+0.1555507+0.1857044+5.721165)
0.1114585/(8.712809+0.1114585+0.03197762+0.1555507+0.1857044+5.721165)
0.03197762/(8.712809+0.1114585+0.03197762+0.1555507+0.1857044+5.721165)
0.1555507/(8.712809+0.1114585+0.03197762+0.1555507+0.1857044+5.721165)
0.1857044/(8.712809+0.1114585+0.03197762+0.1555507+0.1857044+5.721165)
5.721165/(8.712809+0.1114585+0.03197762+0.1555507+0.1857044+5.721165)

deltacost_lt = (tot_costs_lt_post-tot_costs_lt)
deltacost_lt[deltacost_lt>0]=mean(deltacost_lt)
summary(deltacost_lt)
quantile(deltacost_lt,c(.025,.975))
sum(((costs_obese+prod_obese)*prev_obese_post)*alive_post-((costs_obese+prod_obese)*prev_obese)*alive)/sum(tot_costs_lt_post-tot_costs_lt)*100
sum(((costs_chd+prod_chd)*prev_chd_post)*alive_post-((costs_chd+prod_chd)*prev_chd)*alive)/sum(tot_costs_lt_post-tot_costs_lt)*100
sum(((costs_str+prod_str)*prev_str_post)*alive_post-((costs_str+prod_str)*prev_str)*alive)/sum(tot_costs_lt_post-tot_costs_lt)*100
sum(((costs_dm+prod_dm)*prev_dm_post)*alive_post-((costs_dm+prod_dm)*prev_dm)*alive)/sum(tot_costs_lt_post-tot_costs_lt)*100
sum(((costs_ckd+prod_ckd)*prev_ckd_post)*alive_post-((costs_ckd+prod_ckd)*prev_ckd)*alive)/sum(tot_costs_lt_post-tot_costs_lt)*100
sum(((costs_den+prod_den)*prev_den_post)*alive_post-((costs_den+prod_den)*prev_den)*alive)/sum(tot_costs_lt_post-tot_costs_lt)*100
1674.584/(1674.584+1005.32+1624.502+784.2466+2476.855+967.9582)
1005.32/(1674.584+1005.32+1624.502+784.2466+2476.855+967.9582)
1624.502/(1674.584+1005.32+1624.502+784.2466+2476.855+967.9582)
784.2466/(1674.584+1005.32+1624.502+784.2466+2476.855+967.9582)
2476.855/(1674.584+1005.32+1624.502+784.2466+2476.855+967.9582)
967.9582/(1674.584+1005.32+1624.502+784.2466+2476.855+967.9582)


```


# subgroup analyses

```{r sub}

write_csv(as.data.frame(rowMeans(dfsub)),path="meandfsub.csv")
write_csv(as.data.frame(rowSds(dfsub)),path="sdfsub.csv")



```
